Split out the dtrace parser and collapser.
At this point this comes at a performance penalty for the
workflow of pure collapsing. However the split has undergone
no performance work at this point and we are not measuring
the difference of full flamegraph creation that could now save
a whole set of encoding and decoding stack traces.

```
collapse/dtrace         time:   [11.913 ms 12.349 ms 12.843 ms]
                        thrpt:  [102.66 MiB/s 106.77 MiB/s 110.67 MiB/s]
                 change:
                        time:   [+12.176% +15.353% +19.268%] (p = 0.00 < 0.05)
                        thrpt:  [-16.155% -13.309% -10.854%]
                        Performance has regressed.
Found 12 outliers among 100 measurements (12.00%)
  12 (12.00%) high mild
```
